+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to the "htmlcreator" extension will be documented in this fi
 
 ## [Unreleased]
 
-
 ## [1.0.0] - 2021-10-29
 ### Added
 - Publish on MarketPlace
+
+## [1.0.1] - 2021-10-29
+### Added
+- Possibility to add frameworks
+- Add Title input, and edit HTML <title>
+
+### Changed
+- Reformat Code
+- Optimize Code